Big list of changes, see full commit message
Binary Interpreter WPF:
 - Added BioTlkFileSet to parsable classes with full parser support

Bio2DA Editor WPF:
 - Big update with TLK preview, live preview of active cell, optimize memory usage, WIP

Script Editor WPF:
 - Ported over Unhood bytecode parsing to ME3Explorer for ME1. Can now decompile ME1 functions, however it is not yet integrated into our script editor, just outputing the functon text for now. ME1 script support is now best in the toolkit

Package Editor WPF:
 - Add experiments - Build ME1 native function file

Property Collection Engine:
 - Remove ME3 check when finding start of primitive component properties, which now lets ME1/ME2 see primitive component properties
 - ME1 Object Info file has been updated to include transient properties. You can now add these, though it's unknown if they'll do anything, but it fixes a lot of parsing issues ME1 had. Will eventually do ME2.

+++ ME3Explorer/Bio2DAEditor/Bio2DA.cs
@@ -198,7 +198,7 @@ namespace ME3Explorer
                                 int index = (rowindex * ColumnNames.Count()) + colindex; //+1 because they are not zero based indexes since they are numerals
                                 stream.WriteBytes(BitConverter.GetBytes(index));
                             }
-                            stream.WriteByte(cell.Type);
+                            stream.WriteByte((byte)cell.Type);
                             stream.WriteBytes(cell.Data);
                         }
                         else
